fix: Support bigger seed.sql files using docker copy (#295)
* Support bigger seed.sql files using docker copy

For seed.sql files that exceed the maximum argument length of the Linux
distribution used to run the Postgres container the postgres -c command
will fail. Supabase CLI will then silently fail to apply the seed file.

A new Docker utility function DockerAddFile can be used to write a byte
buffer to a container, in this case the seed.sql file. Then migrating
the seed file is just a matter of using the file option of the pgsql
command rather than the command option.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>

* Improve error handling and fix reset command

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>