fix: Propagate scanner errors (#470)
For really long lines, it is possible for the scanner to run out of
buffer space, resulting in

```go
scanner.Err() // bufio.Scanner: token too long
```

This error was not checked however, causing a silent failure which was
supressed.

This commit adds a check for `scanner.Err()` and properly propagates any
non-nil values.