fix: Increase buffer capacity (#470)
For really long lines, it is possible for the scanner to run out of
buffer space, resulting in

```go
scanner.Err() // bufio.Scanner: token too long
```

Following [this](https://stackoverflow.com/a/39864391/1137077)
suggestion from Stack Overflow, this commit implements a workaround
where the maximum buffer capacity is increased from 64K to 256K. Testing
against the SQL file with geodata from #470 showed that already 128K was
enough to resolve that issue. 256K is double that and - as a wise man
once said - "ought to be enough for anybody".