Update GoTrue Image to avoid signup link generation bug
GoTrue version 2.145.0 has a bug where generating a signup link will fail with "user_not_found", which was fixed in 2.146.0 (https://github.com/supabase/auth/pull/1514). Since 2.145.0 is locked in the CLI, we can't get the bug fix out.

This PR updates the image to 2.149.0 which is the most recent version as of 2024-04-22 so the signup link generation works again.