fix: Use discoveryURL when fetching third-party OIDC config
When fetching the well-known configuration for an OIDC third-party auth provider, we used the issuerURL which is just the baseURL instead of the discoveryURL as seems intended. This caused Auth0 configurations to try to parse HTML instead of the correct configuration, and I suspect this is the case for all the other providers as well.

With this bugfix it should work as expected for at least Auth0.