fix: set healthcheck for Mailpit container (#4010)
The mailpit container's built-in healthcheck does not work with podman,
which causes `supabase start` to fail without the
`--ignore-health-check` flag.

I'm not 100% sure why this is, but I suspect it's because the mailpit
[healthcheck has `--start-interval=1s`][1], and the `--start-interval`
option was [only added to Docker Engine since v25.0][2]
(or v1.44 of the Docker API), which podman does not yet suppprt
(see https://github.com/containers/podman/issues/26505).

Setting an explicit healthcheck matches the healthchecks we set for the
other containers, and fixes this issue.

[1]: https://github.com/axllent/mailpit/blob/d26e317d2507015d4d9fa247207ccd6a68fe645f/Dockerfile#L28
[2]: https://docs.docker.com/reference/dockerfile/#healthcheck
Fixes half of: https://github.com/supabase/cli/issues/3566

Co-authored-by: Han Qiao <sweatybridge@gmail.com>